+++ examples/typed_inputs/app.py
@@ -25,7 +25,6 @@ class ShinyInputs(Inputs):
 
 
 def server(input: Inputs, output: Outputs, session: Session):
-
     # Cast `input` to our ShinyInputs class. This just tells the static type checker
     # that we want it treated as a ShinyInputs object for type checking; it has no
     # run-time effect.