+++ shiny/examples/event/app.py
@@ -35,7 +35,6 @@ app_ui = ui.page_fluid(
 
 
 def server(input: Inputs, output: Outputs, session: Session):
-
     # Update a random number every second
     val = reactive.Value(random.randint(0, 1000))