+++ shiny/reactive/_reactives.py
@@ -31,6 +31,7 @@ if TYPE_CHECKING:
 
 T = TypeVar("T")
 
+
 # ==============================================================================
 # Value
 # ==============================================================================
@@ -537,7 +538,9 @@ class Effect_:
             except Exception as e:
                 traceback.print_exc()
 
-                warnings.warn("Error in Effect: " + str(e), ReactiveWarning)
+                warnings.warn(
+                    "Error in Effect: " + str(e), ReactiveWarning, stacklevel=2
+                )
                 if self._session:
                     await self._session._unhandled_error(e)