+++ shiny/session/_session.py
@@ -265,7 +265,9 @@ class Session(object, metaclass=SessionMeta):
                             message, object_hook=_utils.lists_to_tuples
                         )
                     except json.JSONDecodeError:
-                        warnings.warn("ERROR: Invalid JSON message", SessionWarning)
+                        warnings.warn(
+                            "ERROR: Invalid JSON message", SessionWarning, stacklevel=2
+                        )
                         return
 
                     if "method" not in message_obj:
@@ -273,7 +275,6 @@ class Session(object, metaclass=SessionMeta):
                         return
 
                     async with lock():
-
                         if message_obj["method"] == "init":
                             verify_state(ConnectionState.Start)
 
@@ -326,7 +327,7 @@ class Session(object, metaclass=SessionMeta):
                 self._run_session_end_tasks()
 
     def _manage_inputs(self, data: Dict[str, object]) -> None:
-        for (key, val) in data.items():
+        for key, val in data.items():
             keys = key.split(":")
             if len(keys) > 2:
                 raise ValueError(
@@ -404,6 +405,7 @@ class Session(object, metaclass=SessionMeta):
                 warnings.warn(
                     "Received uploadEnd message for non-existent upload operation.",
                     SessionWarning,
+                    stacklevel=2,
                 )
                 return None
             file_data = upload_op.finish()
@@ -463,6 +465,7 @@ class Session(object, metaclass=SessionMeta):
                                     "@session.download(filename=) to specify one "
                                     "manually",
                                     SessionWarning,
+                                    stacklevel=2,
                                 )
                                 filename = download_id
 
@@ -489,7 +492,6 @@ class Session(object, metaclass=SessionMeta):
                         wrapped_contents: AsyncIterable[bytes]
 
                         if isinstance(contents, AsyncIterable):
-
                             # Need to wrap the app-author-provided iterator in a
                             # callback that installs the appropriate context mgrs.
                             # We already use this context mgrs further up in the
@@ -800,7 +802,6 @@ class Session(object, metaclass=SessionMeta):
         return f"session/{urllib.parse.quote(self.id)}/dynamic_route/{urllib.parse.quote(name)}?nonce={urllib.parse.quote(nonce)}"
 
     def _process_ui(self, ui: TagChildArg) -> RenderedDeps:
-
         res = TagList(ui).render()
         deps: List[Dict[str, Any]] = []
         for dep in res["dependencies"]:
@@ -866,11 +867,17 @@ class SessionProxy:
 # Inputs
 # ======================================================================================
 
+
 # TODO: provide a real input typing example when we have an answer for that
 # https://github.com/rstudio/py-shiny/issues/70
 class Inputs:
     """
     A class representing Shiny input values.
+
+    This class provides access to a :class:`~shiny.session.Session`'s input values. The
+    input values are reactive :class:`~shiny.reactive.Values`, and can be accessed with
+    the ``[]`` operator, or with ``.``. For example, if there is an input named ``x``,
+    it can be accessed via ``input["x"]()`` or ``input.x()``.
     """
 
     def __init__(
@@ -913,6 +920,13 @@ class Inputs:
 
     def __delattr__(self, key: str) -> None:
         self.__delitem__(key)
+
+    def __contains__(self, key: str) -> bool:
+        # This looks simple, but does a number of things. By accessing `self[key]`, it
+        # indirectly calls `__getitem__`, which applies a namespace to the key, and
+        # it populates the key if it doesn't exist yet. It then calls `is_set()`, which
+        # creates a reactive dependency, and returns whether the value is set.
+        return self[key].is_set()
 
 
 # ======================================================================================